backend/requirements.txt
numpy<2.0.0
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
python-multipart==0.0.9
openai==1.12.0
llama-index-core==0.10.12
llama-index-vector-stores-chroma==0.1.4
llama-index-embeddings-huggingface==0.1.4
chromadb==0.4.24
sentence-transformers==2.5.1
psutil==5.9.8
pywebview==4.4.1
pyjwt==2.8.0
bcrypt==4.1.2
pypdf==4.0.2
docx2txt==0.8
